Add support for Symfony 3 components
Loosen the version constraint for symfony/css-selector so
the library could be loaded together with Symfony 2 or Symfony 3
components depending on each other.

Symfony CSS selector component deprecated the static interface for converting
from CSS to XPath in 2.8 and removed it in 3.0. This will actually check
which is available and will use the new interface if possible, but it
will still fallback if you have installed a pre-2.8 version of the CSS
selector component.

See https://github.com/OpenBuildings/spiderling/pull/18 for details.

Closes https://github.com/OpenBuildings/spiderling/pull/18.