Do not depend on Request::factory() for initial requests
Kohana behaviour for `Request::factory()` is different depending on
whether an initial request is already created or not.

We can stop depending on it by just using the constructor
and setting the query string with a setter.

`$_GET` is still set for `URL::query()` and other functions
which depend on globals unfortunately.